# docs/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cognitive Quorum API",
    "description": "Backend for Cognitive Quorum application.",
    "version": "2.1.0"
  },
  "paths": {
...
    "/auth/verify": {
      "post": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Verify User Token",
        "description": "Exchanges a Firebase ID Token (or mock token) for the Backend User Profile.\n\nArgs:\n    payload (TokenPayload): The token payload.\n    auth_service (AuthServiceDep): Authentication service dependency.\n\nReturns:\n    LoginResponse: The authenticated user profile and status.\n\nRaises:\n    HTTPException: If the user is found in Firebase but not in the DB (404),\n                   or if the token is invalid (401).",
        "operationId": "verify_user_token_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/impersonate": {
      "post": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Impersonate User",
        "description": "Generates an impersonation token for the target user. Requires ROOT.\n\nArgs:\n    request (ImpersonationRequest): Payload containing target_uid.\n    current_user (CurrentUserDep): The requesting user (must be ROOT).\n    auth_service (AuthServiceDep): Auth service.\n\nReturns:\n    ImpersonationResponse: The access token.\n\nRaises:\n    HTTPException: If permission denied (403) or target not found (404).",
        "operationId": "impersonate_user_auth_impersonate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpersonationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/users": {
      "post": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Create User",
        "description": "Create a new user.\n\nArgs:\n    user_data (UserCreate): Payload for the new user.\n    current_user (CurrentUserDep): The requesting user (must be ROOT, ADMIN, or MANAGER).\n    auth_service (AuthServiceDep): Authentication service dependency.\n\nReturns:\n    User: The created user profile.\n\nRaises:\n    HTTPException: If permission denied (403) or validation fails (400).",
        "operationId": "create_user_auth_users_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "List Users",
        "description": "List users visible to the current user (scoped by Organization).\n\nArgs:\n    current_user (CurrentUserDep): The requesting user.\n    auth_service (AuthServiceDep): Authorization service.\n\nReturns:\n    list[User]: A list of accessible user profiles.",
        "operationId": "list_users_auth_users_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "title": "Response List Users Auth Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/organizations": {
      "post": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Create Organization",
        "description": "Create a new Tenant Organization.\n\nArgs:\n    org_data (OrganizationCreate): Payload for the new organization.\n    current_user (CurrentUserDep): The requesting user (must be ROOT).\n    auth_service (AuthServiceDep): Authentication service dependency.\n\nReturns:\n    Organization: The created organization.\n\nRaises:\n    HTTPException: If user is not ROOT (403).",
        "operationId": "create_organization_auth_organizations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/users/{uid}": {
      "delete": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Delete User",
        "description": "Delete a user.\n\nEnforces Last Admin Protection.\n\nArgs:\n    uid (str): The UID of the user to delete.\n    current_user (CurrentUserDep): The requesting user (ROOT or ADMIN).\n    auth_service (AuthServiceDep): Authorization service.\n\nReturns:\n    dict: Status confirmation.\n\nRaises:\n    HTTPException: Permission denied (403) or business logic error (400).",
        "operationId": "delete_user_auth_users__uid__delete",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Update User",
        "description": "Update a user (Role, Display Name, etc).\n\nArgs:\n    uid (str): The UID of the user to update.\n    user_update (UserUpdate): Fields to update.\n    current_user (CurrentUserDep): Requesting user.\n    auth_service (AuthServiceDep): Authorization service.\n\nReturns:\n    User: The updated user profile.",
        "operationId": "update_user_auth_users__uid__patch",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication & Users"
        ],
        "summary": "Get My Profile",
        "description": "Get the currently authenticated user's profile.\n\nArgs:\n    current_user (CurrentUserDep): The authenticated user.\n    auth_service (AuthServiceDep): Auth service.\n\nReturns:\n    User: The full user profile.",
        "operationId": "get_my_profile_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/organizations/": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Organization",
        "description": "Create a new Tenant Organization.\n\nArgs:\n    org (OrganizationCreateRequest): Organization details.\n    user (TokenData): Requesting user (ROOT required).\n    auth (AuthServiceDep): Authentication service.\n    repo (RepositoryDep): Repository dependency.\n\nReturns:\n    OrganizationResponse: The created organization.\n\nRaises:\n    HTTPException: If ID conflict (409).",
        "operationId": "create_organization_organizations__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Organizations",
        "description": "List all organizations.\n\nArgs:\n    user (TokenData): Requesting user (must be ROOT).\n    repo (RepositoryDep): Repository dependency.\n\nReturns:\n    list[OrganizationResponse]: List of all organizations.",
        "operationId": "list_organizations_organizations__get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationResponse"
                  },
                  "title": "Response List Organizations Organizations  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/me": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get My Organization",
        "description": "Get the organization of the current user.\n\nArgs:\n    user (CurrentUserDep): Requesting user.\n    repo (RepositoryDep): Repository dependency.\n\nReturns:\n    OrganizationResponse: organization details.",
        "operationId": "get_my_organization_organizations_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpersonationRequest": {
        "properties": {
          "target_uid": {
            "type": "string",
            "title": "Target Uid"
          }
        },
        "type": "object",
        "required": [
          "target_uid"
        ],
        "title": "ImpersonationRequest",
        "description": "Request payload for impersonation."
      },
      "ImpersonationResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "ImpersonationResponse",
        "description": "Response containing the impersonation token."
      },
...
      "LoginResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token_valid": {
            "type": "boolean",
            "title": "Token Valid"
          },
          "debug_msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug Msg"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token_valid"
        ],
        "title": "LoginResponse",
        "description": "Response model for successful login.\n\nAttributes:\n    user (User): The full user profile.\n    token_valid (bool): Confirmation of token validity.\n    debug_msg (Optional[str]): Debug information (mock vs real)."
      },
...
      "OrganizationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "admin_email": {
            "type": "string",
            "format": "email",
            "title": "Admin Email"
          },
          "admin_password": {
            "type": "string",
            "title": "Admin Password"
          },
          "admin_name": {
            "type": "string",
            "title": "Admin Name"
          }
        },
        "type": "object",
        "required": [
          "name",
          "admin_email",
          "admin_password",
          "admin_name"
        ],
        "title": "OrganizationCreate",
        "description": "Payload for creating a new organization.\n\nAttributes:\n    name (str): Organization display name.\n    admin_email (EmailStr): Email for the initial admin user.\n    admin_password (str): Password for the initial admin user.\n    admin_name (str): Display name for the initial admin user."
      },
...
      "TokenPayload": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "TokenPayload",
        "description": "Payload for token verification.\n\nAttributes:\n    token (str): The Firebase or Mock ID Token."
      },
...
    }
  }
}